Extracted `CalloutNodeComponent` to separate file
no issue

- no code changes, extracted to match common pattern of separate Node and NodeComponent files
